use BedrockAgentCoreApp from bedrock-agentcore SDK
Replace LangGraph-only server with BedrockAgentCoreApp which is the
actual Bedrock AgentCore SDK. The app exposes /health and /agent routes
(FastAPI-compatible) for test infrastructure compatibility alongside the
standard /invocations AgentCore entrypoint. OneAgent auto-instruments
the underlying boto3 Bedrock API calls.
