+++ aws-bedrock-agents/oneagent/Makefile
@@ -2,8 +2,6 @@ SHELL            := /bin/bash
 -include .env
 export
 
-APP_IMAGE        ?=
-BUILD_PLATFORM   ?= linux/amd64
 PORT             ?= 8000
 
 .PHONY: install run build push request help
@@ -15,13 +13,13 @@ install: ## Install Python dependencies
 	uv sync
 
 run: ## Run app locally on port $(PORT)
-	uv run python3 -m uvicorn server:app --host 0.0.0.0 --port $(PORT)
+	uv run python3 main.py
 
-build: ## Build container image  (APP_IMAGE, BUILD_PLATFORM)
-	docker buildx build --platform $(BUILD_PLATFORM) -t $(APP_IMAGE) .
+build: ## Not applicable — no container defined for this demo
+	@echo "build: no Dockerfile for this demo"
 
-push: ## Build and push container image to registry
-	docker buildx build --no-cache --platform $(BUILD_PLATFORM) -t $(APP_IMAGE) --push .
+push: ## Not applicable — no container defined for this demo
+	@echo "push: no Dockerfile for this demo"
 
 request: ## POST /agent to localhost:$(PORT)
 	curl -s -X POST http://localhost:$(PORT)/agent \
