+++ aws-bedrock-agents/oneagent/README.md
@@ -10,33 +10,34 @@ This example contains a demo of a Personal Assistant Agent built on top of [Bedr
 > For detailed setup instructions, configuration options, and advanced use cases, please refer to the [Get Started Docs](https://docs.dynatrace.com/docs/shortlink/ai-ml-get-started).
 
 Bedrock AgentCore comes with [Observability](https://docs.aws.amazon.com/bedrock-agentcore/latest/devguide/observability.html) support out-of-the-box.
-We just need to register an [OpenTelemetry SDK](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/overview.md#sdk) to send the data to Dynatrace AI Observability.
-
-We simplified this process, hiding all the complexity inside [dynatrace.py](./dynatrace.py).
-For sending data to your Dynatrace tenant, you can configure the `OTEL_ENDPOINT` env var with your Dynatrace URL for ingesting [OTLP](https://docs.dynatrace.com/docs/shortlink/otel-getstarted-otlpexport), for example: `https://wkf10640.live.dynatrace.com/api/v2/otlp`.
-
-The API access token will be read from your filesystem under `/etc/secrets/dynatrace_otel`. 
-
+Dynatrace OneAgent automatically instruments the underlying AWS Bedrock API calls without any code changes required.
 
 ## How to use
 
 ### Setting your AWS keys
 
 Follow the [Amazon Bedrock AgentCore documentation](https://docs.aws.amazon.com/bedrock-agentcore/latest/devguide/runtime-permissions.html) to configure your AWS Role with the correct policies.
-Afterwards, you can set your AWS keys in your environment variables by running the following command in your terminal:
-
+Afterwards, set your AWS credentials in environment variables:
 
 ```bash
-export AWS_ACCESS_KEY_ID==your_api_key
-export AWS_SECRET_ACCESS_KEY==your_secret_key
-export AWS_REGION=your_region
+export AWS_ACCESS_KEY_ID=your_api_key
+export AWS_SECRET_ACCESS_KEY=your_secret_key
+export AWS_DEFAULT_REGION=us-east-1
 ```
 
-Ensure your account has access to the model `us.anthropic.claude-haiku-4-5-20251001-v1:0` used in this example. Please refer to the
-[Amazon Bedrock documentation](https://docs.aws.amazon.com/bedrock/latest/userguide/model-access-permissions.html) to see how to enable access to the model.
-You can change the model used by configuring the environment variable `BEDROCK_MODEL_ID`.
-
+Ensure your account has access to the model `us.anthropic.claude-haiku-4-5-20251001-v1:0` used in this example. Refer to the
+[Amazon Bedrock documentation](https://docs.aws.amazon.com/bedrock/latest/userguide/model-access-permissions.html) to enable model access.
+You can change the model by setting the `BEDROCK_MODEL_ID` environment variable.
 
 ### Run the app
 
-`uv run main.py`
+```bash
+make install
+make run
+```
+
+The agent is available at `http://localhost:8000/agent`. Send a POST request with a `task` field:
+
+```bash
+make request
+```
